backend/requirements.txt
# FastAPI ve Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyPDF2==3.0.1
reportlab==4.0.7
google-generativeai==0.3.1
python-multipart==0.0.6
typing-extensions==4.8.0
